- fixed type of services.goodsla. It's float now. (Alexei)
git-svn-id: svn://svn.zabbix.com/trunk@699 f4e7fd02-d210-4e34-a1e0-d4093fb66cc2

+++ create/postgresql/schema.sql
@@ -384,7 +384,7 @@ CREATE TABLE services (
   algorithm		int2		DEFAULT '0' NOT NULL,
   triggerid		int4,
   showsla		int4		DEFAULT '0' NOT NULL,
-  goodsla		int4		DEFAULT '99.9' NOT NULL,
+  goodsla		float8		DEFAULT '99.9' NOT NULL,
   PRIMARY KEY (serviceid)
 );
